Migrate to entry_points {console_scripts}
This modification allows `ogr` to be executed from the command line on
Windows immeditately after `pip install ...` without any further action
on the user's part (assuming `%pythonhome%\Scripts` is in PATH).

It should have no impact on Linux users. Entry_points is now the
recommended method for cross platform compatibility.[*][1]

It would be possible to get to the same end a little bit more simply by
moving `bin/ogr` to `ogrtools/ogr.py` instead of creating a package out
of it. However I a) didn't want to step into the main package as I
don't understand it yet, and b) as it's own package will make it easier
to use something like [Click][2] for extending the CLI and keep that
churn out of the core.

[1] https://python-packaging-user-guide.readthedocs.org/en/latest/distributing.html#scripts
[2] http://click.pocoo.org/

Also see:
http://stackoverflow.com/questions/30518806/python-migrate-setup-py-scripts-to-entry-points/